+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Added visualizers for the new LOS functions.
 - Made it so challenge moon scores won't be submitted with Imperium enabled.
 - Added the possibility to change the custom welcome message.
+- Night Vision now works outdoors and in orbit too.
+- Night Vision now follows active camera in Spectate, Freecam and any other mode.
 
 ### Bug Fixes